+++ src/exe/chain-sync/Main.hs
@@ -74,10 +74,10 @@ main = do
 
   -- Chain sync loop
 
-  atlasConfig <- Data.Maybe.fromMaybe (error "Atlas configuration failed") <$> decodeConfigEnvOrFile "ATLAS_CORE_CONFIG" defaultAtlasCoreConfig
+  atlasConfig <- fromMaybe (error "Atlas configuration failed") <$> decodeConfigEnvOrFile "ATLAS_CORE_CONFIG" defaultAtlasCoreConfig
   let networkId = cfgNetworkId atlasConfig
 
-  deployedScriptsContext <- Data.Maybe.fromMaybe (error "Deployed validators configuration failed") <$> decodeConfigEnvOrFile @DeployedScriptsContext "DEPLOYED_VALIDATORS_CONFIG" defaultTxBuldingContextFile
+  deployedScriptsContext <- fromMaybe (error "Deployed validators configuration failed") <$> decodeConfigEnvOrFile @DeployedScriptsContext "DEPLOYED_VALIDATORS_CONFIG" defaultTxBuldingContextFile
   let mpHash = getMintingPolicyHash deployedScriptsContext
 
   let policyHexText = T.pack $ printf "%s" mpHash